+++ packages/api-reference/src/components/Content/BaseUrl/index.ts
@@ -1,0 +1,1 @@
+export { default as BaseUrl } from './BaseUrl.vue'
