+++ packages/api-reference/src/components/Content/ClientLibraries/index.ts
@@ -1,0 +1,1 @@
+export { default as ClientLibraries } from './ClientLibraries.vue'
