+++ packages/api-reference/src/components/Content/Models/index.ts
@@ -1,0 +1,2 @@
+export { default as Models } from './Models.vue'
+export { default as ModelsAccordion } from './ModelsAccordion.vue'
