+++ packages/api-reference/src/components/Content/Operation/EndpointDetailsCard.vue
