+++ packages/api-reference/src/components/Content/Operation/EndpointPath.vue
