+++ packages/api-reference/src/components/Content/Operation/OperationAccordion.vue
@@ -5,8 +5,8 @@ import { ScalarIcon, ScalarIconButton } from '@scalar/components'
 import { useClipboard, useNavState } from '../../../hooks'
 import type { Tag, TransformedOperation } from '../../../types'
 import { Anchor } from '../../Anchor'
+import { MarkdownRenderer } from '../../MarkdownRenderer'
 import { SectionAccordion } from '../../Section'
-import MarkdownRenderer from '../MarkdownRenderer.vue'
 import EndpointDetailsCard from './EndpointDetailsCard.vue'
 import EndpointPath from './EndpointPath.vue'
 import ExampleRequest from './ExampleRequest.vue'
