+++ packages/api-reference/src/components/Content/Operation/PathResponses/ExampleResponse.vue
