+++ packages/api-reference/src/components/Content/Operation/PathResponses/Headers.vue
