+++ packages/api-reference/src/components/Content/Operation/PathResponses/RawSchema.vue
