+++ packages/api-reference/src/components/Content/Operation/index.ts
@@ -1,0 +1,2 @@
+export { default as Operation } from './Operation.vue'
+export { default as OperationAccordion } from './OperationAccordion.vue'
