--- packages/api-reference/src/components/Content/ReferenceEndpoint/index.ts
@@ -1,1 +1,0 @@
-export { default as default } from './ReferenceEndpoint.vue'
