+++ packages/api-reference/src/components/Content/Schema/index.ts
@@ -1,0 +1,3 @@
+export { default as Schema } from './Schema.vue'
+export { default as SchemaHeading } from './SchemaHeading.vue'
+export { default as SchemaProperty } from './SchemaProperty.vue'
