+++ packages/api-reference/src/components/Content/Tag/TagAccordion.vue
@@ -1,10 +1,10 @@
 <!-- eslint-disable vue/no-unused-properties -->
 <script setup lang="ts">
-import { useNavState } from '../../hooks'
-import { type Spec, type Tag } from '../../types'
-import { Anchor } from '../Anchor'
-import { SectionContainerAccordion, SectionHeader } from '../Section'
-import MarkdownRenderer from './MarkdownRenderer.vue'
+import { useNavState } from '../../../hooks'
+import type { Spec, Tag } from '../../../types'
+import { Anchor } from '../../Anchor'
+import { MarkdownRenderer } from '../../MarkdownRenderer'
+import { SectionContainerAccordion, SectionHeader } from '../../Section'
 
 defineProps<{
   tag: Tag
