+++ packages/api-reference/src/components/Content/Tag/index.ts
@@ -1,0 +1,2 @@
+export { default as Tag } from './Tag.vue'
+export { default as TagAccordion } from './TagAccordion.vue'
