+++ packages/api-reference/src/components/Content/Webhooks/index.ts
@@ -1,0 +1,1 @@
+export { default as Webhooks } from './Webhooks.vue'
