+++ packages/api-reference/src/components/MarkdownRenderer/index.ts
@@ -1,0 +1,1 @@
+export { default as MarkdownRenderer } from './MarkdownRenderer.vue'
