+++ packages/api-reference/src/components/Sidebar/SidebarGroup.vue
