+++ packages/api-reference/src/components/Sidebar/index.ts
@@ -1,0 +1,1 @@
+export { default as Sidebar } from './Sidebar.vue'
