+++ packages/api-reference/src/index.ts
@@ -5,14 +5,12 @@ export {
 export { default as ApiReference } from './components/ApiReference.vue'
 export { default as ApiReferenceLayout } from './components/ApiReferenceLayout.vue'
 export { default as ApiReferenceBase } from './components/ApiReferenceBase.vue'
-export { default as Sidebar } from './components/Sidebar.vue'
-export { default as SidebarGroup } from './components/SidebarGroup.vue'
-export { default as SidebarElement } from './components/SidebarElement.vue'
 export { default as RenderedReference } from './components/Content/Content.vue'
 export { default as DarkModeToggle } from './components/DarkModeToggle.vue'
 export { default as SearchModal } from './components/SearchModal.vue'
 export { default as SearchButton } from './components/SearchButton.vue'
 
+export * from './components/Sidebar'
 export * from './components/Card'
 
 export * from './stores'
